+++ app/javascript/designer/DesignerDataTab.tsx
@@ -275,6 +275,8 @@ export default function DesignerDataTab({ designStack, setDesignStack }: Designe
             <option value={"axis_hidden_units"}>axis hidden units</option>
             <option value={"allied_random_drop_0"}>allied random drop deploy setup</option>
             <option value={"allied_random_drop_2"}>allied random drop deploy turn 2</option>
+            <option value={"axis_escape"}>axis unit escape</option>
+            <option value={"allied_escape"}>allied unit escape</option>
             <option value={"axis_bonzai"}>axis bonzai</option>
             <option value={"axis_cave_movement"}>axis cave movement</option>
             <option value={"winter"}>winter: no digging in, water as open for infantry</option>